refactor: CookieManager 패키지 변경
