+++ src/main/java/roomescape/common/cookie/CookieManager.java
@@ -1,4 +1,4 @@
-package roomescape.client;
+package roomescape.common.cookie;
 
 import jakarta.servlet.http.Cookie;
 import roomescape.auth.controller.AuthController;
